\name{weighted_sum}
\alias{weighted_sum}
\title{Weighted Sum}
\description{
  Computes the weighted sum of values.
}
\usage{
weighted_sum(values, weights)
}
\arguments{
  \item{values}{Numeric vector of values}
  \item{weights}{Numeric vector of weights}
}
\value{
  Weighted sum.
}
\examples{
weighted_sum(c(1, 2, 3), c(0.2, 0.3, 0.5))  # Returns 2.3
}
\author{
Samir Baladi \email{gitdeeper@gmail.com}
}
